+++ backend/src/meetup/mod.rs
@@ -1,1 +1,4 @@
 pub mod search;
+// TODO: these should not be pub
+pub mod search_request;
+pub mod search_types;
